+++ src/common/data_structures.ts
@@ -32,7 +32,9 @@ export interface Capabilities {
   logsDebug?: boolean;
   multiModalActionPlanning?: boolean;
   narratedActions?: boolean;
+  perceivedLatencyReport?: boolean;
   phonemes?: boolean;
+  pingPongReport?: boolean;
   silence?: boolean;
 }
 
@@ -138,6 +140,7 @@ export enum InworldPacketType {
   CONVERSATION_EVENT = 'CONVERSATION_EVENT',
   ENTITIES_ITEM_OPERATION = 'ENTITIES_ITEM_OPERATION',
   OPERATION_STATUS = 'OPERATION_STATUS',
+  LATENCY_REPORT = 'LATENCY_REPORT',
 }
 
 export enum InworlControlAction {
@@ -155,6 +158,11 @@ export enum InworldConversationEventType {
   STARTED = 'STARTED',
   UPDATED = 'UPDATED',
   EVICTED = 'EVICTED',
+}
+
+export enum InworldLatencyReportType {
+  PERCEIVED_LATENCY = 'PERCEIVED_LATENCY',
+  PING_PONG = 'PING_PONG',
 }
 
 export enum ConversationState {