+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Server to Client log packets
+- Added latency reporting
 
 ## [1.15.0] - 2024-09-17